Fix unused parameters:
 *  src/imodel.C
 *  src/substitution-cache.C
 *  src/sample-branch-lengths.C
 *  src/tools/alignment-draw.C
 *  src/tools/pickout.C
 *  src/alphabet.C
 *  src/alphabet.H
 *  src/moves.C
 *  src/bali-phy.C
 *  src/substitution.C
 *  src/alignment-constraint.C
 *  src/alignment-sums.C
 *  src/dp-matrix.C
 *  src/alignment-constraint.H
 *  src/dp-matrix.H
 *  src/setup-smodel.C
 *  src/smodel.C
 *  src/dp-array.H
 *  src/smodel.H
 *  src/sample-two-nodes.C
 *  src/sample-topology-NNI.C
 *  src/5way.C
 *  src/rng.C
 *  src/exponential.C
 *  src/5way.H
 *  src/refcount.H
 *  src/exponential.H
 *  src/sample-topology-SPR.C

+++ src/substitution-cache.C
@@ -82,7 +82,7 @@ void Multi_Likelihood_Cache::set_length(int t,int l) {
 
   // Increase overall length if necessary
   if (l>C) {
-    int l2 = 4+(1.1*l);
+    int l2 = 4+(int)(1.1*l);
     int delta = l2-C;
     C = l2;
